relocate JoinButton for ClubCard
